El crud funciona

+++ web/python/controlador.py
@@ -1,7 +1,7 @@
 from wsgiref.simple_server import make_server
 from jinja2 import Environment, FileSystemLoader
 import os
-from modelos import get_nadadores, insert_nadador, parse_post_data, redirect_to_main, delete_nadador
+from modelos import get_nadadores, insert_nadador, parse_post_data, redirect_to_main, delete_nadador, update_nadador
 from views import serve_static
 
 
@@ -12,6 +12,7 @@ template = env.get_template('index.html')
 # Define la función app que manejará las solicitudes.
 def app(environ, start_response):
     path = environ.get('PATH_INFO')
+    print(path)
     template
     if path == '/':
         return handle_crud(environ, start_response)
@@ -31,14 +32,17 @@ def handle_crud(environ, start_response):
         start_response(status, response_headers)
         return [response]
     elif environ['REQUEST_METHOD'] == 'POST':
-        # Manejar solicitudes POST para insertar o eliminar datos
+        # Manejar solicitudes POST para insertar, eliminar o actualizar datos
         post_data = parse_post_data(environ)
         if post_data:
             if 'id' in post_data:
                 nadador_id = int(post_data['id'])
-                delete_nadador(nadador_id)
+                if 'action' in post_data and post_data['action'] == 'delete':
+                    delete_nadador(nadador_id)  # Llama a la función para eliminar el registro
+                elif 'action' in post_data and post_data['action'] == 'update':
+                    update_nadador(post_data)  # Llama a la función para actualizar el registro
             else:
-                insert_nadador(post_data)
+                insert_nadador(post_data)  # Llama a la función para insertar el registro
             # Redirigir al usuario a la página principal después de la acción
             return redirect_to_main(environ, start_response)
     else:
